+++ docs/release-notes.md
@@ -19,6 +19,7 @@ Minor changes:
 - ProxmoxVE: Define DNS entries for every interface, not just the first
 - Hetzner: Add the `HETZNER_PUBLIC_IPV6` attribute
 - ProxmoxVE: Explicitely deactivate Dracut IP autoconf as the IP is statically set
+- Oracle Cloud: Handle nested metadata values for instance pool members
 
 Packaging changes: